requirements.txt
natsort
argparse
google.cloud
six
